add progression jauge
- add display function for a jauge to show up on the right side of the screen and indicating player and lava progression through the map height
- remove lava distance counter
- refactor arrays initialisation by removing number type flag
- update mushroom sprite
- decrease maximum obstacle spawn rate

+++ src/model/dataclasses.py
@@ -22,33 +22,33 @@ class Obstacle:
 
 @dataclass(frozen=True)
 class ObstacleParticle:
-    pos: ndarray
+    pos: ndarray  # in world space
     timer: float = 0
 
 
 @dataclass(frozen=True)
 class GooParticle:
-    pos: ndarray
+    pos: ndarray  # in world space
     velocity: ndarray
     radius: float = GOO_PARTICLE_INIT_RADIUS
 
 
 @dataclass(frozen=True)
 class Player:
-    pos: ndarray
+    pos: ndarray  # in world space
     rect: Rect
-    velocity: ndarray = zeros(2, dtype=float)
+    velocity: ndarray = zeros(2)
     on_ground: bool = False
     obstacle_collision: bool = False
 
 
 @dataclass(frozen=True)
 class Grapple:
-    start: ndarray = zeros(2, dtype=float)  # start position of the grapple (=player position) in world space
-    end: ndarray = zeros(2, dtype=float)  # end position of the grapple in world space
-    head: ndarray = zeros(2, dtype=float)
-    head_velocity: ndarray = zeros(2, dtype=float)
-    head_start: ndarray = zeros(2, dtype=float)  # point at which the head is fired
+    start: ndarray = zeros(2)  # start position of the grapple (=player position) in world space
+    end: ndarray = zeros(2)  # end position of the grapple in world space
+    head: ndarray = zeros(2)
+    head_velocity: ndarray = zeros(2)
+    head_start: ndarray = zeros(2)  # point at which the head is fired
 
 
 @dataclass(frozen=True)
@@ -60,6 +60,6 @@ class Lava:
 @dataclass(frozen=True)
 class Camera:
     center: ndarray  # world position on which the camera is centered
-    heading: ndarray = zeros(2, dtype=float)  # world vector pointing on focus position
-    top_left: ndarray = zeros(2, dtype=float)  # world position of the top left corner of the camera
-    offset: ndarray = zeros(2, dtype=float)  # offset to apply to a sprite position to get its screen position
+    heading: ndarray = zeros(2)  # world vector pointing on focus position
+    top_left: ndarray = zeros(2)  # world position of the top left corner of the camera
+    offset: ndarray = zeros(2)  # offset to apply to a sprite position to get its screen position
